fix tests for move /rpc/tendermint in tests
