+++ rpc/tendermint/client/client.go
@@ -4,7 +4,7 @@ import (
 	rpcclient "github.com/tendermint/go-rpc/client"
 
 	acm "github.com/eris-ltd/eris-db/account"
-	ctypes "github.com/eris-ltd/eris-db/rpc/core/types"
+	ctypes "github.com/eris-ltd/eris-db/rpc/tendermint/core/types"
 	"github.com/eris-ltd/eris-db/txs"
 )
