+++ rpc/tendermint/core/accounts.go
@@ -3,7 +3,7 @@ package core
 import (
 	"fmt"
 	acm "github.com/eris-ltd/eris-db/account"
-	ctypes "github.com/eris-ltd/eris-db/rpc/core/types"
+	ctypes "github.com/eris-ltd/eris-db/rpc/tendermint/core/types"
 	. "github.com/tendermint/go-common"
 )
