+++ rpc/tendermint/core/net.go
@@ -1,7 +1,7 @@
 package core
 
 import (
-	ctypes "github.com/eris-ltd/eris-db/rpc/core/types"
+	ctypes "github.com/eris-ltd/eris-db/rpc/tendermint/core/types"
 	sm "github.com/eris-ltd/eris-db/manager/eris-mint/state"
 	"github.com/tendermint/tendermint/types"
